Backend/requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-dotenv